PGF: Fix inconsistent options for \documentclass
This could cause a mismatch between lengths when measuring vs.
producing output for things that depend on the font size setting at
the beginning of the document, e.g. unicode-math.

Use \documentclass{article} without options here, which defaults to
a font size of 10pt (same as the matplotlib rc default for
font.size).